v7.8
+ Odroids | Gracefully handle cases where Odroids have no /boot/boot.ini, e.g. base on upstream kernel and switch to more compatible G_CONFIG_INJECT to apply settings.

+++ PREP_SYSTEM_FOR_DIETPI.sh
@@ -567,11 +567,11 @@ Currently installed: $G_DISTRO_NAME (ID: $G_DISTRO)"; then
 			G_EXEC mv "DietPi-$G_GITBRANCH/boot_xu4.ini" /boot/boot.ini
 			G_EXEC sed -i "s/root=UUID=[^[:blank:]]*/root=UUID=$(findmnt -Ufnro UUID -M /)/" /boot/boot.ini
 
-		elif (( $G_HW_MODEL == 12 )); then
+		elif [[ $G_HW_MODEL == 12 && -f '/boot/boot.ini' ]]; then
 
 			G_EXEC mv "DietPi-$G_GITBRANCH/boot_c2.ini" /boot/boot.ini
 
-		elif (( $G_HW_MODEL == 15 )); then
+		elif [[ $G_HW_MODEL == 15 && -f '/boot/boot.ini' ]]; then
 
 			G_EXEC mv "DietPi-$G_GITBRANCH/boot_n2.ini" /boot/boot.ini
 			G_EXEC sed -i "s/root=UUID=[^[:blank:]]*/root=UUID=$(findmnt -Ufnro UUID -M /)/" /boot/boot.ini
